ml/requirements.txt

# ============================================================================
# FundN3xus ML & RAG - Complete Dependencies (Python 3.13 Compatible)
# ============================================================================
# This file combines all requirements for:
# - ML API Server (server.py)
# - RAG API Server (rag_server.py)
# - Training scripts (train_model.py, train_rag_embeddings.py)
# - Optional: Pinecone cloud vector database
#
# Install: pip install -r requirements.txt
# ============================================================================

# ----------------------------------------------------------------------------
# CORE ML & DATA SCIENCE (for server.py)
# ----------------------------------------------------------------------------
scikit-learn>=1.5.0
xgboost>=2.1.0
imbalanced-learn>=0.12.0
joblib>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
scipy>=1.11.0

# ----------------------------------------------------------------------------
# WEB FRAMEWORK & API (for both servers)
# ----------------------------------------------------------------------------
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# ----------------------------------------------------------------------------
# RAG FRAMEWORK (for rag_server.py)
# ----------------------------------------------------------------------------
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0

# ----------------------------------------------------------------------------
# VECTOR DATABASE (for RAG)
# ----------------------------------------------------------------------------
chromadb>=0.4.22           # Local vector DB (default)
# faiss-cpu>=1.7.4         # Alternative: Facebook's similarity search

# ----------------------------------------------------------------------------
# EMBEDDINGS (for RAG)
# ----------------------------------------------------------------------------
sentence-transformers>=2.2.2   # Local embeddings (default)

# ----------------------------------------------------------------------------
# DOCUMENT PROCESSING (for RAG)
# ----------------------------------------------------------------------------
pypdf>=3.17.0              # PDF processing
python-docx>=1.1.0         # Word documents
openpyxl>=3.1.0           # Excel files
beautifulsoup4>=4.12.0    # HTML parsing

# ----------------------------------------------------------------------------
# UTILITIES (for RAG)
# ----------------------------------------------------------------------------
tiktoken>=0.5.0           # Token counting
tenacity>=8.2.0           # Retry logic

# ----------------------------------------------------------------------------
# OPTIONAL: PINECONE CLOUD VECTOR DATABASE
# ----------------------------------------------------------------------------
# Uncomment these if you want to use Pinecone instead of ChromaDB:
# pinecone-client>=2.2.4
# pinecone-text>=0.7.0

# ----------------------------------------------------------------------------
# OPTIONAL: MONITORING & LOGGING
# ----------------------------------------------------------------------------
# python-json-logger>=2.0.0     # Structured logging
# prometheus-client>=0.19.0     # Metrics

# ----------------------------------------------------------------------------
# OPTIONAL: DEVELOPMENT TOOLS
# ----------------------------------------------------------------------------
# pytest>=7.4.0                 # Testing framework
# black>=23.0.0                 # Code formatting
# flake8>=6.0.0                 # Code linting
# mypy>=1.5.0                   # Type checking

# ----------------------------------------------------------------------------
# OPTIONAL: DATA VISUALIZATION
# ----------------------------------------------------------------------------
# matplotlib>=3.8.0
# seaborn>=0.13.0

# ----------------------------------------------------------------------------
# OPTIONAL: GPU SUPPORT
# ----------------------------------------------------------------------------
# Uncomment if you have CUDA-enabled GPU:
# nvidia-ml-py3>=7.352.0
# cupy-cuda11x>=12.0.0         # Adjust CUDA version as needed
